Move winit/wasm example to example directory
