+++ sg2d-wgpu/Cargo.toml
@@ -6,6 +6,11 @@ edition = "2021"
 [lib]
 crate-type = ["cdylib", "rlib"]
 
+[[example]]
+name = "wgpu-winit"
+path = "examples/wgpu-winit"
+crate-type = ["cdylib", "rlib"]
+
 [dependencies]
 sg2d = { path = "../sg2d" }
 
@@ -24,22 +29,8 @@ futures-intrusive = "^0.5"
 glyphon = { git = "https://github.com/grovesNL/glyphon.git", rev="941309aed230d7110bfec0d4af502ecb4648cf90" }
 lyon = { workspace = true }
 
-# Only used for lib.rs example. Move this to example
+[dev-dependencies]
 sg2d-vega = { path = "../sg2d-vega" }
 serde_json = { version = "1.0.111" }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-console_error_panic_hook = "0.1.6"
-console_log = "1.0"
-wgpu = { version = "0.18", features = ["webgl"]}
-wasm-bindgen = "=0.2.87"
-wasm-bindgen-futures = "0.4.30"
-web-sys = { version = "0.3.53", features = [
-    "Document",
-    "Window",
-    "Element",
-]}
-
-[dev-dependencies]
 dssim = "3.2.4"
 rstest = "0.18.2"
