+++ sg2d-wgpu/examples/wgpu-winit/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "wgpu-winit"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+# Empty workspace needed so cargo doesn't try to add example as a workspace member
+[workspace]
+
+[dependencies]
+sg2d-vega = { path = "../../../sg2d-vega" }
+sg2d-wgpu = { path = "../../../sg2d-wgpu" }
+sg2d = { path = "../../../sg2d" }
+
+cfg-if = "1"
+winit = "0.28"
+env_logger = "0.10"
+log = "0.4"
+wgpu = "0.18"
+pollster = "0.3"
+serde_json = { version = "1.0.111" }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+console_error_panic_hook = "0.1.6"
+console_log = "1.0"
+wgpu = { version = "0.18", features = ["webgl"]}
+wasm-bindgen = "=0.2.87"
+wasm-bindgen-futures = "0.4.30"
+web-sys = { version = "0.3.53", features = [
+    "Document",
+    "Window",
+    "Element",
+]}
