+++ sg2d/Cargo.toml
@@ -8,4 +8,4 @@ edition = "2021"
 [dependencies]
 thiserror = { workspace = true }
 serde = { workspace = true }
-lyon_path = { workspace = true }
+lyon_path = { workspace = true, features = ["serialization"] }
